add map, youtube video, link, and rendering
